Allow for configurable disabling of the automatic pageView event by the JavaScript.
