+++ src/main/resources/divolte.js
@@ -34,6 +34,8 @@ var COOKIE_DOMAIN = '';
 var LOGGING = false;
 /** @define {string} */
 var SCRIPT_NAME = 'divolte.js';
+/** @define {boolean} */
+var AUTO_PAGE_VIEW_EVENT = true;
 
 (function (global, factory) {
   factory(global);
@@ -845,13 +847,9 @@ var SCRIPT_NAME = 'divolte.js';
                   parameterValue;
               switch (customParameterType) {
                 case 'string':
-                  parameterValue = customParameter;
-                  break;
                 case 'number':
-                  parameterValue = customParameter.toString(36);
-                  break;
                 case 'boolean':
-                  parameterValue = customParameter ? 't' : 'f';
+                  parameterValue = customParameter.toString();
                   break;
                 default:
                   throw "Parameter '" + customName + "' is of unsupported type: " + customParameterType;
@@ -940,19 +938,23 @@ var SCRIPT_NAME = 'divolte.js';
       visibilityEventName = "webkitvisibilitychange";
     }
 
+    var signalPageView = AUTO_PAGE_VIEW_EVENT ? function() {
+      signal('pageView');
+    } : function() {};
+
     if (typeof hiddenProperty !== 'undefined' && document[hiddenProperty]) {
       // The {add|remove}EventListener functions are not available in <= IE8;
       // but this branch shouldn't execute in that case, since the hidden
       // property is also undefined.
       document.addEventListener(visibilityEventName, function visibilityListener() {
         if (document[hiddenProperty] === false) {
-          signal('pageView');
+          signalPageView();
           document.removeEventListener(visibilityEventName, visibilityListener);
         }
       })
     } else {
       // TODO: Possibly defer until the DOM is ready?
-      signal('pageView');
+      signalPageView();
     }
   } else {
     warn("Divolte module already initialized; existing module left intact.");
