+++ src/main/resources/reference.conf
@@ -147,6 +147,12 @@ divolte {
     // When true, the served JavaScript will be compiled, but not
     // minified, improving readability when debugging in the browser.
     debug = false
+    
+    // When false, divolte.js will not automatically send a pageView
+    // event after being loaded. This way clients can send a initial
+    // event themselves and have full control over the event type and
+    // the custom parameters that are sent with the initial event. 
+    auto_page_view_event = true
   }
 
   // This section controls settings related to the processing of incoming
