+++ src/test/resources/static/quirks/test-basic-page.html
@@ -26,6 +26,6 @@
   </head>
   <body>
     <script src="/divolte.js" defer async></script>
-    <div id="custom" onclick="divolte.signal('custom',{ key: 'value' })">Click me to fire an event!</div>
+    <div id="custom" onclick="divolte.signal('custom',{ key: 'value', answer: 42, magic: true })">Click me to fire an event!</div>
   </body>
 </html>
