+++ src/test/resources/static/quirks/test-custom-page-view.html
@@ -1,6 +1,5 @@
-<!DOCTYPE html>
 <!--
-  ~ Copyright 2014 GoDataDriven B.V.
+  ~ Copyright 2015 GoDataDriven B.V.
   ~
   ~ Licensed under the Apache License, Version 2.0 (the "License");
   ~ you may not use this file except in compliance with the License.
@@ -26,7 +25,10 @@
     <link rel="icon" href="data:;base64,iVBORw0KGgo=">
   </head>
   <body>
-    <script src="/divolte.js" defer async></script>
-    <div id="custom" onclick="divolte.signal('custom',{ key: 'value', answer: 42, magic: true })">Click me to fire an event!</div>
+    <!-- not defered, as we need the module right away -->
+    <script src="/divolte.js"></script>
+    <script>
+      divolte.signal('pageView', { foo: "moo", bar: 'baz' })
+    </script>
   </body>
 </html>
